Handle backup write errors and improve timer cancel (#1165)
* Handle backup write errors and improve timer cancel

Added error handling for backup file write failures in porter/__init__.py to prevent crashes when backup mode is not enabled. Updated MonitorCron.cancel in run/helper.py to join the timer thread after cancellation for better thread management.

* Fix monitor condition and update test epochs

Corrected the logic for enabling system monitoring in SwanLabRun by changing the condition from 'or' to 'and'. Increased the number of epochs in the test_sync unit test to 10000 for more extensive testing.